refactor: extract event_push and imou_life_device_key helper
